pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bili-summary"
version = "0.1.0"
description = "Bilibili video subtitle extraction and summarization tool"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "yt-dlp>=2023.0",
    "dashscope>=1.15.0",
    "qrcode>=7.4.2",
]

[project.scripts]
bili-summary = "bili_summary.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "pytest-httpx>=0.35.0",
    "black>=24.0",
    "isort>=5.0",
    "mypy>=1.0",
    "ruff>=0.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
